+++ src/arkiver/manifest-validator.ts
@@ -1,10 +1,5 @@
-import {
-  assertEquals,
-  assertExists,
-} from 'https://deno.land/std@0.192.0/testing/asserts.ts'
 import { supportedChains } from '../chains.ts'
 import { scope } from '../deps.ts'
-import { Manifest } from './manifest-builder/manifest.ts'
 
 export const parseArkiveManifest = scope({
   manifest: {
@@ -50,18 +45,3 @@ export const parseArkiveManifest = scope({
     handler: 'Function',
   },
 }).compile()
-
-Deno.test('parseArkiveManifest', () => {
-  const manifestBuilder = new Manifest('test')
-
-  manifestBuilder.addChain('ethereum', (chain) => {
-    chain.addContract([]).addSources({ '*': 1n })
-  })
-
-  const manifest = manifestBuilder.build()
-
-  const { problems, data } = parseArkiveManifest.manifest(manifest)
-
-  assertEquals(problems, undefined)
-  assertExists(data)
-})
